# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateRoomRequest:
    properties:
      color:
        type: string
      home_id:
        type: integer
      icon:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Diploma project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Diploma Server API
  version: "1.0"
paths:
...
  /homes/{home_id}/notifications:
    get:
      description: Get all notifications for a home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get notifications by home ID
      tags:
      - notification
  /homes/{home_id}/notifications/{notification_id}:
    delete:
      description: Mark a home notification as read
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark home notification as read
      tags:
      - notification
...
  /homes/{home_id}/rooms:
    get:
      description: Get all rooms in a home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get rooms by home ID
      tags:
      - room
    post:
      consumes:
      - application/json
      description: Create a new room in a home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Create Room Request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new room
      tags:
      - room
  /homes/{home_id}/rooms/{room_id}:
    delete:
      description: Delete a room by ID
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete room
      tags:
      - room
    get:
      description: Get room details by ID
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get room by ID
      tags:
      - room
...
  /homes/notifications:
    get:
      description: Get notifications for the current user, optionally scoped to one
        home via ?home_id=
      parameters:
      - description: Home ID to scope notifications to
        in: query
        name: home_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get notifications by user ID
      tags:
      - notification
  /homes/notifications/{notification_id}:
    delete:
      description: Mark a notification as read
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - notification
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"